+++ tests/nv/ex/test_scanner.py
@@ -559,6 +559,7 @@ class Test_scan_command(unittest.TestCase):
         self.assertRoute(['setlocal opt=val', 'setl opt=val'], cmd('setlocal', params={'option': 'opt', 'value': 'val'}))  # noqa: E501
         self.assertRoute(['shell', 'sh'], cmd('shell'))
         self.assertRoute(['silent ls', 'sil ls'], cmd('silent', params={'command': 'ls'}))  # noqa: E501
+        self.assertRoute(['silent! ls', 'sil! ls'], cmd('silent', params={'command': 'ls'}, forced=True))  # noqa: E501
         self.assertRoute(['snoremap abc xyz', 'snor abc xyz'], cmd('snoremap', params={'lhs': 'abc', 'rhs': 'xyz'}))  # noqa: E501
         self.assertRoute(['snoremap', 'snor'], cmd('snoremap'))
         self.assertRoute(['sort i', 'sor i'], cmd('sort', params={'options': 'i'}, addressable=True))
@@ -566,6 +567,8 @@ class Test_scan_command(unittest.TestCase):
         self.assertRoute(['sort u', 'sor u'], cmd('sort', params={'options': 'u'}, addressable=True))
         self.assertRoute(['sort ui', 'sor ui'], cmd('sort', params={'options': 'ui'}, addressable=True))
         self.assertRoute(['sort', 'sor'], cmd('sort', addressable=True))
+        self.assertRoute(['spellgood fizz', 'spe fizz'], cmd('spellgood', params={'word': 'fizz'}))
+        self.assertRoute(['spellundo fizz', 'spellu fizz'], cmd('spellundo', params={'word': 'fizz'}))
         self.assertRoute(['split file.txt', 'sp file.txt'], cmd('split', params={'file': 'file.txt'}))
         self.assertRoute(['split', 'sp'], cmd('split'))
         self.assertRoute(['substitute', 's'], cmd('substitute', addressable=True))